+++ pyincore/analyses/mlenabledcgejoplin/__init__.py
@@ -4,4 +4,4 @@
 # terms of the Mozilla Public License v2.0 which accompanies this distribution,
 # and is available at https://www.mozilla.org/en-US/MPL/2.0/
 
-from pyincore.analyses.mlenabledcgejoplin.mlcgejoplin import MLEnabledCgeJoplin
+from pyincore.analyses.mlenabledcgejoplin.mlcgejoplin import MlEnabledCgeJoplin