+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+- MlEnabledCgeJoplin analysis with documentation [#588](https://github.com/IN-CORE/pyincore/issues/588)
+
 ## [1.19.0] - 2024-06-12
 
 ### Changed